Fixed test/example dependencies
