+++ example/CMakeLists.txt
@@ -3,6 +3,8 @@
 # Examples
 #
 
+find_package(OpenCV)
+
 if(${ION_BB_BUILD_base})
     ion_compile(producerx3_compile SRCS producerx3_compile.cc PIPELINE_NAME producerx3)
     ion_run(producerx3 producerx3_compile SRCS producerx3_run.cc TARGET_STRING "host-profile")
@@ -13,35 +15,33 @@ endif()
 
 if(${ION_BB_BUILD_image-processing} AND ${ION_BB_BUILD_image-io} AND ${ION_BB_BUILD_sgm})
     ion_compile(sgm_compile SRCS sgm_compile.cc PIPELINE_NAME sgm)
-    ion_run(sgm sgm_compile SRCS sgm_run.cc TARGET_STRING "host-profile")
+    ion_run(sgm sgm_compile SRCS sgm_run.cc INCS ${OpenCV_INCLUDE_DIR} LIBS ion-bb ${OpenCV_LIBRARIES} TARGET_STRING "host-profile")
     if(${CUDA_FOUND})
-        ion_run(sgm_gpu sgm_compile SRCS sgm_run.cc TARGET_STRING "host-cuda-cuda_capability_50-profile")
+        ion_run(sgm_gpu sgm_compile SRCS sgm_run.cc INCS ${OpenCV_INCLUDE_DIR} LIBS ion-bb ${OpenCV_LIBRARIES} TARGET_STRING "host-cuda-cuda_capability_50-profile")
     endif()
 endif()
 
 if(${ION_BB_BUILD_base} AND ${ION_BB_BUILD_image-io} AND ${ION_BB_BUILD_dnn})
-    find_package(OpenCV)
-
     ion_compile(dnn_compile SRCS dnn_compile.cc PIPELINE_NAME dnn)
-    ion_run(dnn dnn_compile SRCS dnn_run.cc TARGET_STRING "host-profile")
+    ion_run(dnn dnn_compile SRCS dnn_run.cc INCS ${OpenCV_INCLUDE_DIR} LIBS ion-bb ${OpenCV_LIBRARIES} TARGET_STRING "host-profile")
     if(${CUDA_FOUND})
-        ion_run(dnn_gpu dnn_compile SRCS dnn_run.cc TARGET_STRING "host-cuda-cuda_capability_50-profile")
+        ion_run(dnn_gpu dnn_compile SRCS dnn_run.cc INCS ${OpenCV_INCLUDE_DIR} LIBS ion-bb ${OpenCV_LIBRARIES} TARGET_STRING "host-cuda-cuda_capability_50-profile")
     endif()
     ion_jit(dnn_jit SRCS dnn.cc INCS ${OpenCV_INCLUDE_DIRS} LIBS ${OpenCV_LIBRARIES})
 endif()
 
 if(${ION_BB_BUILD_base} AND ${ION_BB_BUILD_image-processing} AND ${ION_BB_BUILD_image-io})
     ion_compile(demo_compile SRCS demo_compile.cc PIPELINE_NAME demo)
-    ion_run(demo demo_compile SRCS demo_run.cc TARGET_STRING "host-profile")
+    ion_run(demo demo_compile SRCS demo_run.cc INCS ${OpenCV_INCLUDE_DIR} LIBS ion-bb ${OpenCV_LIBRARIES} TARGET_STRING "host-profile")
     if(${CUDA_FOUND})
-        ion_run(demo_gpu demo_compile SRCS demo_run.cc TARGET_STRING "host-cuda-cuda_capability_50-profile")
+        ion_run(demo_gpu demo_compile SRCS demo_run.cc INCS ${OpenCV_INCLUDE_DIR} LIBS ion-bb ${OpenCV_LIBRARIES} TARGET_STRING "host-cuda-cuda_capability_50-profile")
     endif()
 
-    ion_jit(demo_jit SRCS demo.cc)
+    ion_jit(demo_jit SRCS demo.cc INCS ${OpenCV_INCLUDE_DIR} LIBS ion-bb ${OpenCV_LIBRARIES})
 endif()
 
 if(${ION_BB_BUILD_image-processing} AND ${ION_BB_BUILD_image-io})
-    ion_jit(isp_jit SRCS isp.cc)
+    ion_jit(isp_jit SRCS isp.cc INCS ${OpenCV_INCLUDE_DIR} LIBS ${OpenCV_LIBRARIES})
     ion_register_test(isp isp_jit TARGET_STRING "host-profile")
     if(${CUDA_FOUND}})
         ion_register_test(isp_gpu isp_jit TARGET_STRING "host-cuda-cuda_capability_50-profile")
@@ -49,7 +49,7 @@ if(${ION_BB_BUILD_image-processing} AND ${ION_BB_BUILD_image-io})
 endif()
 
 if(${ION_BB_BUILD_image-processing} AND ${ION_BB_BUILD_image-io} AND ${ION_BB_BUILD_sgm})
-    ion_jit(isp_and_sgm_jit SRCS isp_and_sgm.cc)
+    ion_jit(isp_and_sgm_jit SRCS isp_and_sgm.cc INCS ${OpenCV_INCLUDE_DIR} LIBS ${OpenCV_LIBRARIES})
     ion_register_test(isp_and_sgm isp_and_sgm_jit TARGET_STRING "host-profile")
     if(${CUDA_FOUND}})
         # Temporaly disabled
@@ -74,7 +74,7 @@ if(${ION_BB_BUILD_image-io} AND UNIX AND NOT APPLE)
     ion_jit(v4l2_jit SRCS v4l2.cc)
     ion_register_test(v4l2 v4l2_jit TARGET_STRING "host-profile")
 
-    ion_jit(realsense_jit SRCS realsense.cc)
+    ion_jit(realsense_jit SRCS realsense.cc INCS ${OpenCV_INCLUDE_DIR} LIBS ${OpenCV_LIBRARIES})
     ion_register_test(realsense realsense_jit TARGET_STRING "host-profile")
 endif()
